Split up pkg/input/record_reader.go
