+++ pkg/input/constants.go
@@ -1,0 +1,3 @@
+package input
+
+const CSV_BOM = "\xef\xbb\xbf"
